+++ lib/src/file-system.browser.ts
@@ -1,3 +1,7 @@
-export async function readJSON(folder: string, filename: string): Promise<any> {
+export async function readGrammar(langId: string): Promise<any> {
+  throw new Error("no fs");
+}
+
+export async function readTheme(themeName: string): Promise<any> {
   throw new Error("no fs");
 }