+++ lib/src/network.browser.ts
@@ -5,6 +5,13 @@ const LIGHTER_VERSION = "__LIGHTER_VERSION__";
 // /grammars/${name}.json
 // /themes/${name}.json
 export async function fetchJSON(endpoint: string) {
+  console.warn(
+    `Code Hike warning: Fetching resource from network "${endpoint}"`
+  );
   const r = await fetch(`https://lighter.codehike.org/${endpoint}.json`);
+  // console.log(`https://lighter.codehike.org/${endpoint}.json`, r.status);
+  if (!r.ok) {
+    return undefined;
+  }
   return await r.json();
 }